move BriefcaseIdValue to common
