+++ core/backend/src/perftest/PolymorphicQuery.test.ts
@@ -6,9 +6,9 @@ import { assert } from "chai";
 import * as path from "path";
 import { DbResult, Id64, Id64String } from "@bentley/bentleyjs-core";
 import { Arc3d, IModelJson as GeomJson, Point3d } from "@bentley/geometry-core";
-import { Code, ColorDef, GeometricElementProps, GeometryStreamProps, IModel, SubCategoryAppearance } from "@bentley/imodeljs-common";
+import { BriefcaseIdValue, Code, ColorDef, GeometricElementProps, GeometryStreamProps, IModel, SubCategoryAppearance } from "@bentley/imodeljs-common";
 import { Reporter } from "@bentley/perf-tools/lib/Reporter";
-import { BackendRequestContext, BriefcaseIdValue, ECSqlStatement, IModelDb, IModelJsFs, SnapshotDb, SpatialCategory } from "../imodeljs-backend";
+import { BackendRequestContext, ECSqlStatement, IModelDb, IModelJsFs, SnapshotDb, SpatialCategory } from "../imodeljs-backend";
 import { IModelTestUtils } from "../test/IModelTestUtils";
 import { KnownTestLocations } from "../test/KnownTestLocations";
 
@@ -84,8 +84,8 @@ describe("SchemaDesignPerf Polymorphic query", () => {
           else
             baseClassName = `Child${(i - 1).toString()}`;
           sxml = `${sxml}<ECEntityClass typeName="Child${i.toString()}" >
-        <BaseClass>${  baseClassName}</BaseClass>
-              <ECProperty propertyName = "PropChild${  i.toString()}" typeName = "string" />
+        <BaseClass>${baseClassName}</BaseClass>
+              <ECProperty propertyName = "PropChild${i.toString()}" typeName = "string" />
         </ECEntityClass>`;
         }
