+++ core/backend/src/perftest/RelationshipImpact.test.ts
@@ -6,11 +6,15 @@ import { assert } from "chai";
 import * as path from "path";
 import { DbResult, Id64, Id64String } from "@bentley/bentleyjs-core";
 import { Arc3d, IModelJson as GeomJson, Point3d } from "@bentley/geometry-core";
-import { Code, ColorDef, GeometricElementProps, GeometryStreamProps, IModel, RelatedElement, SubCategoryAppearance } from "@bentley/imodeljs-common";
-import { Reporter } from "@bentley/perf-tools/lib/Reporter";
 import {
-  BackendRequestContext, BriefcaseIdValue, ECSqlStatement, IModelDb, IModelJsFs, RelationshipProps, SnapshotDb, SpatialCategory,
-} from "../imodeljs-backend";
+  BriefcaseIdValue, Code, ColorDef, GeometricElementProps, GeometryStreamProps, IModel, RelatedElement, RelationshipProps, SubCategoryAppearance,
+} from "@bentley/imodeljs-common";
+import { Reporter } from "@bentley/perf-tools/lib/Reporter";
+import { BackendRequestContext } from "../BackendRequestContext";
+import { SpatialCategory } from "../Category";
+import { ECSqlStatement } from "../ECSqlStatement";
+import { IModelDb, SnapshotDb } from "../IModelDb";
+import { IModelJsFs } from "../IModelJsFs";
 import { IModelTestUtils } from "../test/IModelTestUtils";
 import { KnownTestLocations } from "../test/KnownTestLocations";
 import { PerfTestUtility } from "./PerfTestUtils";
