+++ core/backend/src/test/standalone/ElementRoundTrip.test.ts
@@ -6,9 +6,10 @@ import { assert, expect } from "chai";
 import { DbResult, Id64, Id64String } from "@bentley/bentleyjs-core";
 import { Arc3d, Cone, IModelJson as GeomJson, Point2d, Point3d } from "@bentley/geometry-core";
 import {
+  BriefcaseIdValue,
   Code, ColorDef, ElementAspectProps, GeometricElementProps, GeometryStreamProps, IModel, SubCategoryAppearance,
 } from "@bentley/imodeljs-common";
-import { BackendRequestContext, BriefcaseIdValue, ECSqlStatement, IModelDb, IModelJsFs, SnapshotDb, SpatialCategory } from "../../imodeljs-backend";
+import { BackendRequestContext, ECSqlStatement, IModelDb, IModelJsFs, SnapshotDb, SpatialCategory } from "../../imodeljs-backend";
 import { ElementRefersToElements } from "../../Relationship";
 import { IModelTestUtils } from "../IModelTestUtils";
